+++ WCF-Generator/App.config
@@ -227,6 +227,11 @@
   <responseDtoGenerator enabled ="true">
     <responseDtoConfigurations>
       <add generatedDtosNameSpace="blah" projectForGeneratedCode="TestResponseDtoGeneration">
+        <prefixStrings>
+          <add text="using System;"/>
+          <add text="using System.Linq;"/>
+          <add text="using TestMappingGeneration.Enums;"/>
+        </prefixStrings>
       </add>
     </responseDtoConfigurations>
   </responseDtoGenerator>
