+++ WCF-Generator/MappingsGeneration/Configuration/MappingConfiguration.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Configuration;
 using System.Linq;
 using WCFGenerator.Common;
+using WCFGenerator.Common.Configuration;
 
 namespace WCFGenerator.MappingsGeneration.Configuration
 {
@@ -123,34 +124,6 @@ namespace WCFGenerator.MappingsGeneration.Configuration
         public PrefixStrings PrefixStrings
         {
             get { return ((PrefixStrings)(base["prefixStrings"])); }
-        }
-    }
-
-    [ConfigurationCollection(typeof(PrefixString))]
-    public class PrefixStrings : ConfigurationElementCollection
-    {
-        protected override ConfigurationElement CreateNewElement()
-        {
-            return new PrefixString();
-        }
-
-        protected override object GetElementKey(ConfigurationElement element)
-        {
-            return ((PrefixString)(element)).Text;
-        }
-
-        public PrefixString this[int idx]
-        {
-            get { return (PrefixString)BaseGet(idx); }
-        }
-    }
-
-    public class PrefixString : ConfigurationElement
-    {
-        [ConfigurationProperty("text", DefaultValue = "", IsKey = true, IsRequired = true)]
-        public string Text
-        {
-            get { return ((string)(base["text"])); }
         }
     }
