+++ WCF-Generator/ResponseDtoGeneration/ResponseDtoGenerator.cs
@@ -4,6 +4,8 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using WCFGenerator.Common;
+using WCFGenerator.Common.Configuration;
+using WCFGenerator.Common.Infrastructure;
 using WCFGenerator.MappingsGeneration.Configuration;
 using WCFGenerator.MappingsGenerator.Analysis;
 using WCFGenerator.ResponseDtoGeneration.Analyser;
@@ -27,6 +29,25 @@ namespace WCFGenerator.ResponseDtoGeneration
         {
             var analyser = new ResponseDtoGenerationAnalyser(_configuration, _generatorWorkspace);
             await analyser.Run();
+            var code = GetFullCode(analyser.Classes);
+            _generatorWorkspace.SetTargetProject(_configuration.ProjectForGeneratedCode);
+            _generatorWorkspace.UpdateFileInTargetProject("GenerateResponseDto.g.cs", "Generation", code);
+            await _generatorWorkspace.ApplyTargetProjectChanges(true);
+        }
+
+        private string GetFullCode(ClassCompilerInfo[] classes)
+        {
+            var sb = new StringBuilder();
+            foreach (PrefixString prefixString in _configuration.PrefixStrings)
+            {
+                sb.AppendLine(prefixString.Text);
+            }
+
+            sb.AppendLine("");
+            sb.AppendLine("namespace " + _configuration.MapExtensionNameSpace);
+            sb.AppendLine("{");
+            sb.AppendLine("}");
+            return sb.ToString();
         }
     }
 }
